Move vorticitytransport_grid to c_sw (#179)
