+++ fv3core/stencils/c_sw.py
@@ -3,7 +3,6 @@ from gt4py.gtscript import __INLINED, PARALLEL, computation, interval, parallel,
 
 import fv3core._config as spec
 import fv3core.stencils.d2a2c_vect as d2a2c
-import fv3core.stencils.vorticitytransport_cgrid as vorticity_transport
 import fv3core.utils.gt4py_utils as utils
 from fv3core.decorators import gtstencil
 from fv3core.utils.corners import fill_4corners_x, fill_4corners_y
@@ -232,6 +231,97 @@ def update_vorticity_and_kinetic_energy(
         ke = 0.5 * dt2 * (ua * ke + va * vort)
 
 
+@gtstencil()
+def update_zonal_velocity(
+    vorticity: sd,
+    ke: sd,
+    velocity: sd,
+    velocity_c: sd,
+    cosa: sd,
+    sina: sd,
+    rdxc: sd,
+    dt2: float,
+):
+    from __externals__ import i_end, i_start, namelist
+
+    with computation(PARALLEL), interval(...):
+        assert __INLINED(namelist.grid_type < 3)
+        # additional assumption: not __INLINED(spec.grid.nested)
+
+        tmp_flux = dt2 * (velocity - velocity_c * cosa) / sina
+        with parallel(region[i_start, :], region[i_end + 1, :]):
+            tmp_flux = dt2 * velocity
+
+        flux = vorticity[0, 0, 0] if tmp_flux > 0.0 else vorticity[0, 1, 0]
+        velocity_c = velocity_c + tmp_flux * flux + rdxc * (ke[-1, 0, 0] - ke)
+
+
+@gtstencil()
+def update_meridional_velocity(
+    vorticity: sd,
+    ke: sd,
+    velocity: sd,
+    velocity_c: sd,
+    cosa: sd,
+    sina: sd,
+    rdyc: sd,
+    dt2: float,
+):
+    from __externals__ import j_end, j_start, namelist
+
+    with computation(PARALLEL), interval(...):
+        assert __INLINED(namelist.grid_type < 3)
+        # additional assumption: not __INLINED(spec.grid.nested)
+
+        tmp_flux = dt2 * (velocity - velocity_c * cosa) / sina
+        with parallel(region[:, j_start], region[:, j_end + 1]):
+            tmp_flux = dt2 * velocity
+
+        flux = vorticity[0, 0, 0] if tmp_flux > 0.0 else vorticity[1, 0, 0]
+        velocity_c = velocity_c - tmp_flux * flux + rdyc * (ke[0, -1, 0] - ke)
+
+
+def vorticitytransport_cgrid(
+    uc: sd, vc: sd, vort_c: sd, ke_c: sd, v: sd, u: sd, dt2: float
+):
+    """Update the C-Grid zonal and meridional velocity fields.
+
+    Args:
+        uc: x-velocity on C-grid (input, output)
+        vc: y-velocity on C-grid (input, output)
+        vort_c: Vorticity on C-grid (input)
+        ke_c: kinetic energy on C-grid (input)
+        v: y-velocity on D-grid (input)
+        u: x-velocity on D-grid (input)
+        dt2: timestep (input)
+    """
+    grid = spec.grid
+    update_meridional_velocity(
+        vort_c,
+        ke_c,
+        u,
+        vc,
+        grid.cosa_v,
+        grid.sina_v,
+        grid.rdyc,
+        dt2,
+        origin=grid.compute_origin(),
+        domain=grid.domain_shape_compute_buffer_2d(add=(0, 1, 0)),
+    )
+    update_zonal_velocity(
+        vort_c,
+        ke_c,
+        v,
+        uc,
+        grid.cosa_u,
+        grid.sina_u,
+        grid.rdxc,
+        dt2,
+        origin=grid.compute_origin(),
+        domain=grid.domain_shape_compute_buffer_2d(add=(1, 0, 0)),
+    )
+
+
 def compute(delp, pt, u, v, w, uc, vc, ua, va, ut, vt, divgd, omga, dt2):
     grid = spec.grid
     dord4 = True
@@ -340,5 +430,5 @@ def compute(delp, pt, u, v, w, uc, vc, ua, va, ut, vt, divgd, omga, dt2):
         origin=grid.compute_origin(),
         domain=(grid.nic + 1, grid.njc + 1, grid.npz),
     )
-    vorticity_transport.compute(uc, vc, vort, ke, v, u, dt2)
+    vorticitytransport_cgrid(uc, vc, vort, ke, v, u, dt2)
     return delpc, ptc
