+++ tests/translate/translate_c_sw.py
@@ -203,3 +203,29 @@ class TranslateKE_C_SW(TranslateFortranData2Py):
             domain=(self.grid.nic + 2, self.grid.njc + 2, self.grid.npz),
         )
         return self.slice_output(inputs, {"ke_c": ke, "vort_c": vort})
+
+
+class TranslateVorticityTransport_Cgrid(TranslateFortranData2Py):
+    def __init__(self, grid):
+        super().__init__(grid)
+        self.compute_func = c_sw.vorticitytransport_cgrid
+        self.in_vars["data_vars"] = {
+            "uc": {},
+            "vc": {},
+            "vort_c": {
+                "istart": grid.is_ - 1,
+                "iend": grid.ie + 1,
+                "jstart": grid.js - 1,
+                "jend": grid.je + 1,
+            },
+            "ke_c": {
+                "istart": grid.is_ - 1,
+                "iend": grid.ie + 1,
+                "jstart": grid.js - 1,
+                "jend": grid.je + 1,
+            },
+            "u": {},
+            "v": {},
+        }
+        self.in_vars["parameters"] = ["dt2"]
+        self.out_vars = {"uc": grid.x3d_domain_dict(), "vc": grid.y3d_domain_dict()}
