--- tests/translate/translate_vorticitytransport_cgrid.py
@@ -1,29 +1,0 @@
-import fv3core.stencils.vorticitytransport_cgrid as VorticityTransport_Cgrid
-
-from .translate import TranslateFortranData2Py
-
-
-class TranslateVorticityTransport_Cgrid(TranslateFortranData2Py):
-    def __init__(self, grid):
-        super().__init__(grid)
-        self.compute_func = VorticityTransport_Cgrid.compute
-        self.in_vars["data_vars"] = {
-            "uc": {},
-            "vc": {},
-            "vort_c": {
-                "istart": grid.is_ - 1,
-                "iend": grid.ie + 1,
-                "jstart": grid.js - 1,
-                "jend": grid.je + 1,
-            },
-            "ke_c": {
-                "istart": grid.is_ - 1,
-                "iend": grid.ie + 1,
-                "jstart": grid.js - 1,
-                "jend": grid.je + 1,
-            },
-            "u": {},
-            "v": {},
-        }
-        self.in_vars["parameters"] = ["dt2"]
-        self.out_vars = {"uc": grid.x3d_domain_dict(), "vc": grid.y3d_domain_dict()}
